requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
pandas==2.3.2
numpy==2.3.3
matplotlib==3.7.1
psycopg2-binary
requests
beautifulsoup4
